Move the common SQL queries for UTXO chain balances into a separate module
