+++ lib/sanbase/clickhouse/historical_balance/fetchers/btc_balance.ex
@@ -15,6 +15,7 @@ defmodule Sanbase.Clickhouse.HistoricalBalance.BtcBalance do
   use Ecto.Schema
 
   import Sanbase.Clickhouse.HistoricalBalance.Utils
+  import Sanbase.Clickhouse.HistoricalBalance.UtxoSqlQueries
 
   require Sanbase.ClickhouseRepo, as: ClickhouseRepo
 
@@ -33,7 +34,7 @@ defmodule Sanbase.Clickhouse.HistoricalBalance.BtcBalance do
 
   @impl Sanbase.Clickhouse.HistoricalBalance.Behaviour
   def assets_held_by_address(address) do
-    {query, args} = current_balance_query(address)
+    {query, args} = current_balance_query(@table, address)
 
     ClickhouseRepo.query_transform(query, args, fn [value] ->
       %{
@@ -57,7 +58,7 @@ defmodule Sanbase.Clickhouse.HistoricalBalance.BtcBalance do
   @impl Sanbase.Clickhouse.HistoricalBalance.Behaviour
   def historical_balance(address, currency, decimals, from, to, interval)
       when is_binary(address) do
-    {query, args} = historical_balance_query(address, from, to, interval)
+    {query, args} = historical_balance_query(@table, address, from, to, interval)
 
     ClickhouseRepo.query_transform(query, args, fn [dt, balance, has_changed] ->
       %{
@@ -76,7 +77,7 @@ defmodule Sanbase.Clickhouse.HistoricalBalance.BtcBalance do
   @impl Sanbase.Clickhouse.HistoricalBalance.Behaviour
   def balance_change(address_or_addresses, _currency, _decimals, from, to)
       when is_binary(address_or_addresses) or is_list(address_or_addresses) do
-    {query, args} = balance_change_query(address_or_addresses, from, to)
+    {query, args} = balance_change_query(@table, address_or_addresses, from, to)
 
     ClickhouseRepo.query_transform(query, args, fn [address, start_balance, end_balance, change] ->
       {address, {start_balance, end_balance, change}}
@@ -88,7 +89,8 @@ defmodule Sanbase.Clickhouse.HistoricalBalance.BtcBalance do
 
   def historical_balance_change(address_or_addresses, _currency, _decimals, from, to, interval)
       when is_binary(address_or_addresses) or is_list(address_or_addresses) do
-    {query, args} = historical_balance_change_query(address_or_addresses, from, to, interval)
+    {query, args} =
+      historical_balance_change_query(@table, address_or_addresses, from, to, interval)
 
     ClickhouseRepo.query_transform(query, args, fn [dt, change] ->
       %{
@@ -100,132 +102,12 @@ defmodule Sanbase.Clickhouse.HistoricalBalance.BtcBalance do
 
   @impl Sanbase.Clickhouse.HistoricalBalance.Behaviour
   def last_balance_before(address, _currency, _decimals, datetime) do
-    query = """
-    SELECT balance
-    FROM #{@table}
-    PREWHERE
-      address = ?1 AND
-      dt <=toDateTime(?2)
-    ORDER BY dt DESC
-    LIMIT 1
-    """
-
-    args = [address, DateTime.to_unix(datetime)]
+    {query, args} = last_balance_before_query(@table, address, datetime)
 
     case ClickhouseRepo.query_transform(query, args, & &1) do
       {:ok, [[balance]]} -> {:ok, balance}
       {:ok, []} -> {:ok, 0}
       {:error, error} -> {:error, error}
     end
-  end
-
-  # Private functions
-
-  defp current_balance_query(address) do
-    query = """
-    SELECT balance
-    FROM #{@table}
-    PREWHERE
-      address = ?1 AND
-    ORDER BY dt DESC
-    LIMIT 1
-    """
-
-    args = [address]
-    {query, args}
-  end
-
-  defp historical_balance_query(address, from, to, interval) when is_binary(address) do
-    interval = Sanbase.DateTimeUtils.str_to_sec(interval)
-    from_unix = DateTime.to_unix(from)
-    to_unix = DateTime.to_unix(to)
-    span = div(to_unix - from_unix, interval) |> max(1)
-
-    query = """
-    SELECT time, SUM(balance), toUInt8(SUM(has_changed))
-      FROM (
-        SELECT
-          toUnixTimestamp(intDiv(toUInt32(?4 + number * ?1), ?1) * ?1) AS time,
-          toFloat64(0) AS balance,
-          toInt8(0) AS has_changed
-        FROM numbers(?2)
-
-      UNION ALL
-
-      SELECT
-        toUnixTimestamp(intDiv(toUInt32(dt), ?1) * ?1) AS time,
-        argMax(balance, dt) AS balance,
-        toUInt8(1) AS has_changed
-      FROM #{@table}
-      PREWHERE
-        address = ?3 AND
-        dt >= toDateTime(?4) AND
-        dt <= toDateTime(?5)
-      GROUP BY time
-    )
-    GROUP BY time
-    ORDER BY time
-    """
-
-    args = [interval, span, address, from_unix, to_unix]
-    {query, args}
-  end
-
-  defp balance_change_query(address_or_addresses, from, to) do
-    addresses = address_or_addresses |> List.wrap() |> List.flatten()
-
-    query = """
-    SELECT
-      address,
-      argMaxIf(balance, dt, dt <= ?3) AS start_balance,
-      argMaxIf(balance, dt, dt <= ?4) AS end_balance,
-      end_balance - start_balance AS diff
-    FROM #{@table} FINAL
-    PREWHERE
-      address IN (?1)
-    GROUP BY address
-    """
-
-    args = [addresses, from, to]
-
-    {query, args}
-  end
-
-  defp historical_balance_change_query(address_or_addresses, from, to, interval) do
-    addresses = address_or_addresses |> List.wrap() |> List.flatten()
-
-    interval = Sanbase.DateTimeUtils.str_to_sec(interval)
-    to_unix = DateTime.to_unix(to)
-    from_unix = DateTime.to_unix(from)
-    span = div(to_unix - from_unix, interval) |> max(1)
-
-    # The balances table is like a stack. For each balance change there is a record
-    # with sign = -1 that is the old balance and with sign = 1 which is the new balance
-    query = """
-    SELECT time, SUM(change)
-      FROM (
-        SELECT
-          toUnixTimestamp(intDiv(toUInt32(?4 + number * ?1), ?1) * ?1) AS time,
-          toFloat64(0) AS change
-        FROM numbers(?2)
-
-      UNION ALL
-
-      SELECT
-        toUnixTimestamp(intDiv(toUInt32(dt), ?1) * ?1) AS time,
-        balance - oldBalance AS change
-      FROM #{@table} FINAL
-      PREWHERE
-        address in (?3) AND
-        dt >= toDateTime(?4) AND
-        dt < toDateTime(?5)
-      GROUP BY address
-    )
-    GROUP BY time
-    ORDER BY time
-    """
-
-    args = [interval, span, addresses, from_unix, to_unix]
-    {query, args}
   end
 end
